requirements.txt
google-adk>=0.1.0
litellm>=0.18.0
newsapi-python>=0.2.7
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.3.0
pydantic
adktools
feedparser